requirements.txt
markdown
pymdown-extensions
python-dotenv